+++ handler_testing_test.go
@@ -3,6 +3,8 @@ package emperror_test
 import (
 	"testing"
 
+	"errors"
+
 	. "github.com/goph/emperror"
 	"github.com/stretchr/testify/assert"
 )
@@ -10,8 +12,8 @@ import (
 func TestTestHandler_Handle(t *testing.T) {
 	handler := new(TestHandler)
 
-	err1 := New("error 1")
-	err2 := New("error 2")
+	err1 := errors.New("error 1")
+	err2 := errors.New("error 2")
 
 	handler.Handle(err1)
 	handler.Handle(err2)
@@ -25,8 +27,8 @@ func TestTestHandler_Handle(t *testing.T) {
 func TestTestHandler_Last(t *testing.T) {
 	handler := new(TestHandler)
 
-	err1 := New("error 1")
-	err2 := New("error 2")
+	err1 := errors.New("error 1")
+	err2 := errors.New("error 2")
 
 	handler.Handle(err1)
 	handler.Handle(err2)