+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.6.0] - 2017-10-26
+
 ### Added
 
 - HttpError interface for representing errors with an HTTP error context
@@ -107,7 +110,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [Bugsnag](https://bugsnag.com) handler
 
 
-[Unreleased]: https://github.com/goph/emperror/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/goph/emperror/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/goph/emperror/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/goph/emperror/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/goph/emperror/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/goph/emperror/compare/v0.2.2...v0.3.0